% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gameboy.R
\name{annotate_pogo}
\alias{annotate_pogo}
\title{Pogo Team Logo as background}
\usage{
annotate_pogo(team = "mystic", opacity = 0.5)
}
\arguments{
\item{team}{string. either 'mystic', 'valor' or 'instinct'}

\item{opacity}{opacity of team logo.}
}
\description{
enhance
}
