% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthbar.R
\name{gghealth}
\alias{gghealth}
\title{healthbar plot}
\usage{
gghealth(data, names, values, sort.names = TRUE, init.size = 3,
  cuts = c(0.8, 0.4, 0.2))
}
\arguments{
\item{data}{data.frame}

\item{names}{column name of names}

\item{values}{column name of values to plot}

\item{sort.names}{order or not}

\item{init.size}{size of the healthbar}

\item{cuts}{where to change the colors}
}
\description{
plot stuff and things
}
