% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_pal.R
\name{poke_pal}
\alias{poke_pal}
\title{Pokemon inspired Palettes}
\usage{
poke_pal(name, n)
}
\arguments{
\item{name}{string. Should be}

\item{n}{integer specifying the number of colors to return.}
}
\value{
A color palette
}
\description{
stuff about things
}
\details{
name matching
}
