% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_pie.R
\name{poke_pie}
\alias{poke_pie}
\title{Poke Pie Charts}
\usage{
poke_pie(path_to_sprites = "", poke)
}
\arguments{
\item{path_to_sprites}{path to sprite files. See Details}

\item{poke}{name or pokedex number of Pokemon}
}
\description{
Create Pie Charts of the color distribution of Pokemon sprites.
}
\details{
Sprites are not provided and have to be download prior of usage. It is important
that the sprites are numbered from 1.png to 801.png. \url{https://github.com/PokeAPI/sprites}
contains all sprites in the right format and has been used for testing.
}
