% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\title{Pokemon Data}
\format{A data frame containing various stats from 801 pokemon. Use \code{str(pokemon)}
for more details.}
\usage{
pokemon
}
\description{
Pokemon Data
}
\references{
Pokémon data download from [Kaggle](https://www.kaggle.com/rounakbanik/pokemon), originally
scraped from [serebii.net](http://serebii.net/)
}
\keyword{datasets}
