% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rokemon.R
\docType{package}
\name{Rokemon}
\alias{Rokemon}
\alias{Rokemon-package}
\title{Rokemon}
\description{
An R package for your favorite game. More Help can be found on \url{https://github.com/schochastics/Rokemon}{github}
}
