% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gameboy.R
\name{scale_color_poketype}
\alias{scale_color_poketype}
\alias{scale_fill_poketype}
\title{Pokemon color scale}
\usage{
scale_color_poketype(...)

scale_fill_poketype(...)
}
\arguments{
\item{...}{common discrete scale parameters for ggplot}
}
\description{
This allows to use the Pokemon type color scale in ggplot.
}
\examples{
library(ggplot2)

ggplot(pokemon,aes(defense,attack))+
  geom_point(aes(col=type1))+
  scale_color_poketype()

ggplot(pokemon,aes(type1))+
  geom_bar(aes(fill=type1))+
  scale_color_poketype()
}
