% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gameboy.R
\name{theme_gameboy}
\alias{theme_gameboy}
\title{Gamyboy inspired ggplot theme}
\usage{
theme_gameboy(title_style = "8Bit")
}
\arguments{
\item{title_style}{string. Either \code{"8Bit"} for *Bit Font or \code{"poke"} for Pokemon title Font.}
}
\description{
Theme based on the first generation gameboy.
}
