% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gameboy.R
\name{theme_gba}
\alias{theme_gba}
\title{Gamyboy Advanced inspired ggplot theme}
\usage{
theme_gba()
}
\description{
Theme based on the Gameboy Advanced.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(pokemon,aes(attack,defense))+
 geom_point(shape = 15,col = c("#006400"),size=2)+
 theme_gameboy()+
 labs(title = "Classic Gameboy Theme")
}
}
