% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gameboy.R
\name{annotate_pogo}
\alias{annotate_pogo}
\title{Pokemon Go Team Logos}
\usage{
annotate_pogo(team = "mystic", opacity = 0.5)
}
\arguments{
\item{team}{string. either 'mystic', 'valor' or 'instinct'}

\item{opacity}{opacity of team logo.}
}
\description{
Enhance your plots with a background picture of your favorite Pokemon Go team.
}
\examples{
\dontrun{
library(tiduverse)

pokemon \%>\%
 dplyr::filter(type1=="water") \%>\%
 ggplot(aes(defense,attack))+
 annotate_pogo(team = "mystic")+
 geom_point()+
 theme_mystic()+
 labs(title="Team Mystic",subtitle="Water Pokemon")
}
}
