% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poke_pal.R
\name{poke_pal}
\alias{poke_pal}
\title{Pokemon inspired Palettes}
\usage{
poke_pal(name, n)
}
\arguments{
\item{name}{string. name of a Pokemon who's colors should be used}

\item{n}{integer specifying the number of colors to return.}
}
\value{
A color palette
}
\description{
Color palettes generated from Pokemon sprites.
}
\details{
All palettes were constructed automatically and might thus not be very meaningful.
Better pokemon palettes can be found in the package \href{https://github.com/timcdlucas/palettetown}{palettetown}
}
\examples{
poke_pal("pikachu",3)
}
