% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gameboy.R
\name{theme_gameboy}
\alias{theme_gameboy}
\title{Gamyboy inspired ggplot theme}
\usage{
theme_gameboy()
}
\description{
Theme based on the first generation of the Gameboy.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(pokemon,aes(attack,defense))+
 geom_point(shape = 15,col = c("#006400"),size=2)+
 theme_gameboy()+
 labs(title = "Classic Gameboy Theme")
}
}
